backend/requirements.txt
# Python dependencies for SOVD Command WebApp - Backend
# Production dependencies with pinned versions (Task I1.T6)

# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database - PostgreSQL ORM and migrations
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0

# Redis client for caching and session management
redis>=5.0.0
slowapi>=0.1.9

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data validation and settings
pydantic>=2.4.0
pydantic-settings>=2.0.0
jsonschema>=4.20.0

# File upload support
python-multipart>=0.0.6

# Structured logging
structlog>=23.2.0

# Monitoring and metrics
prometheus-fastapi-instrumentator>=6.1.0
